v7

+++ Magnus.Application/Dtos/Requests/TransferWarehouseItemRequest.cs
@@ -4,4 +4,4 @@ public record TransferWarehouseItemRequest(
     Guid ProductId,
     int ProductInternalCode,
     string ProductName,
-    decimal Amount);
+    int Amount);
